Allows usage of different RDF-Versions
Different cameras store the RDF-Metadata in different ways. There seem
to be two ways to store RDF-Data by either specivying the value in an
attribute or by adding a text-node. Currently only the attribute-type is
supported and this Commit allows usage of the Text-node-tpe.

The text-node is for example used by the Theta-Camera by Ricoh